Refactored to utilize system functions instead of classes
Keeping with the idea that if you have a fixed set of 'things' (namely objects), as you code evolves, then you primarily want to use more of a functional approach and just add new operations to these things. This effectively makes it so you can add new functions (or systems in this case) to existing data without having to touch other systems. Have begun this idea with a functional implementation of the rendersystem.