Refactored systems to utilize scene copies instead of references
Trying to implement the idea of having no side effects in my major system functions I'm currently taking and return copies of the scene object and fighters in order to making things easier to test and debug. The one big draw back right now is all these copies are going to slow down the game. However, I can optimize later and this way things would be more parrellizable in the future if necessary.