Refactored code to reflect more of a game engine first approach
By 'game engine first' I don't mean I'm trying to create a generalized game engine. I just mean that I'm developing things with the mindset of the game engine being the exe and containing main.cpp and game logic being thorwn in. This is instead of creating the engine code as a series of separate libraries for which my game calls in and utilizes.