Fixed graphics logic causing bug : Fixed animation logic and physics logic preventing animations from displaying properly
Basically, for the first bug which was causing the second character to be drawn at the same position as the first character and had them falling below the speified y height, I needed to add an opengl vetexAttrib call for position before bind. Without doing so the engine just took the first fighter's vertexattib position which was the last bound. Also, noticed an animation logic error where a higher number of animations were shown to be in the animationMap then was actually created. Also had to zero out the y velocity when fighter was on the ground due to gravity continually subtracting from velocity.